requirements.txt
crewai==0.86.0
crewai-tools==0.17.0
google-generativeai==0.8.3
langchain-google-genai==2.0.7
tavily-python==0.5.0
streamlit==1.41.1
reportlab==4.2.5
python-dotenv==1.0.1
